Refactor architecture (#1936)
* move datasource classes to dao module

* fix send4LetterWord bug
