+++ dolphinscheduler-dao/src/main/java/org/apache/dolphinscheduler/dao/datasource/DB2ServerDataSource.java
@@ -14,7 +14,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package org.apache.dolphinscheduler.common.job.db;
+package org.apache.dolphinscheduler.dao.datasource;
 
 import org.apache.dolphinscheduler.common.Constants;
 import org.apache.dolphinscheduler.common.utils.StringUtils;
@@ -26,10 +26,10 @@ import java.sql.DriverManager;
 import java.sql.SQLException;
 
 /**
- * data source of SQL Server
+ * data source of DB2 Server
  */
-public class SQLServerDataSource extends BaseDataSource {
-    private static final Logger logger = LoggerFactory.getLogger(SQLServerDataSource.class);
+public class DB2ServerDataSource extends BaseDataSource {
+    private static final Logger logger = LoggerFactory.getLogger(DB2ServerDataSource.class);
 
     /**
      * gets the JDBC url for the data source connection
@@ -38,12 +38,15 @@ public class SQLServerDataSource extends BaseDataSource {
     @Override
     public String getJdbcUrl() {
         String jdbcUrl = getAddress();
-        jdbcUrl += ";databaseName=" + getDatabase();
-
-        if (StringUtils.isNotEmpty(getOther())) {
-            jdbcUrl += ";" + getOther();
+        if (jdbcUrl.lastIndexOf("/") != (jdbcUrl.length() - 1)) {
+            jdbcUrl += "/";
         }
 
+        jdbcUrl += getDatabase();
+
+        if (StringUtils.isNotEmpty(getOther())) {
+            jdbcUrl += ":" + getOther();
+        }
         return jdbcUrl;
     }
 
@@ -55,14 +58,14 @@ public class SQLServerDataSource extends BaseDataSource {
     public void isConnectable() throws Exception {
         Connection con = null;
         try {
-            Class.forName(Constants.COM_SQLSERVER_JDBC_DRIVER);
+            Class.forName(Constants.COM_DB2_JDBC_DRIVER);
             con = DriverManager.getConnection(getJdbcUrl(), getUser(), getPassword());
         } finally {
             if (con != null) {
                 try {
                     con.close();
                 } catch (SQLException e) {
-                    logger.error("SQL Server datasource try conn close conn error", e);
+                    logger.error("DB2 Server datasource try conn close conn error", e);
                 }
             }
         }
