+++ dolphinscheduler-dao/src/main/java/org/apache/dolphinscheduler/dao/datasource/PostgreDataSource.java
@@ -14,7 +14,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package org.apache.dolphinscheduler.common.job.db;
+package org.apache.dolphinscheduler.dao.datasource;
 
 import org.apache.dolphinscheduler.common.Constants;
 import org.apache.dolphinscheduler.common.utils.StringUtils;
@@ -26,11 +26,12 @@ import java.sql.DriverManager;
 import java.sql.SQLException;
 
 /**
- * data source of spark
+ * data source of postgreSQL
  */
-public class SparkDataSource extends BaseDataSource {
+public class PostgreDataSource extends BaseDataSource {
 
-  private static final Logger logger = LoggerFactory.getLogger(SparkDataSource.class);
+  private static final Logger logger = LoggerFactory.getLogger(PostgreDataSource.class);
+
 
   /**
    * gets the JDBC url for the data source connection
@@ -45,12 +46,8 @@ public class SparkDataSource extends BaseDataSource {
 
     jdbcUrl += getDatabase();
 
-    if (StringUtils.isNotEmpty(getPrincipal())){
-      jdbcUrl += ";principal=" + getPrincipal();
-    }
-
     if (StringUtils.isNotEmpty(getOther())) {
-      jdbcUrl += ";" + getOther();
+      jdbcUrl += "?" + getOther();
     }
 
     return jdbcUrl;
@@ -64,14 +61,14 @@ public class SparkDataSource extends BaseDataSource {
   public void isConnectable() throws Exception {
     Connection con = null;
     try {
-      Class.forName(Constants.ORG_APACHE_HIVE_JDBC_HIVE_DRIVER);
-      con = DriverManager.getConnection(getJdbcUrl(), getUser(), "");
+      Class.forName(Constants.ORG_POSTGRESQL_DRIVER);
+      con = DriverManager.getConnection(getJdbcUrl(), getUser(), getPassword());
     } finally {
       if (con != null) {
         try {
           con.close();
         } catch (SQLException e) {
-          logger.error("Spark datasource try conn close conn error", e);
+          logger.error("Postgre datasource try conn close conn error", e);
         }
       }
     }
